+++ CHANGELOG.md
@@ -27,6 +27,7 @@
   * Updated ImGui to 1.71: [#1362](https://github.com/dartsim/dart/pull/1362)
   * Fixed refresh of LineSegmentShapeNode: [#1381](https://github.com/dartsim/dart/pull/1381)
   * Fixed OSG transparent object sorting: [#1414](https://github.com/dartsim/dart/pull/1414)
+  * Added modifier key support to ImGuiHandler: [#1436](https://github.com/dartsim/dart/pull/1436)
 
 * Parser